feat: Add article about using NRQL for data aggregation in New Relic
This commit adds a new article titled "SQLならGROUP BYしてHAVINGだけど、New RelicのNRQLならFACETしてgetField()するって話". The article explains how to use NRQL to aggregate data in New Relic and includes examples and tips for using the NRQL syntax. It also provides a step-by-step guide on how to achieve similar results as SQL's GROUP BY and HAVING clauses in NRQL. The article is published on the "aeonpeople" publication and is currently set to "false" for publication.

[Issue #X]